refactor: reorganize for better workspace organization and granularity (move examples, pull out providers)
