--- packages/cli/tsconfig.json
@@ -1,14 +1,0 @@
-{
-  "extends": "../../tsconfig.json",
-  "include": ["src/**/*"],
-  "files": ["src/lib.ts", "src/ellmers.ts"],
-  "exclude": ["**/*.test.ts"],
-  "compilerOptions": {
-    "outDir": "dist",
-    "baseUrl": "./src",
-    "rootDir": "./src",
-    "paths": {
-      "#/*": ["./src/*"]
-    }
-  }
-}
