+++ packages/core/package.json
@@ -5,24 +5,31 @@
   "description": "Ellmers is a tool for building and running DAG pipelines of AI tasks.",
   "scripts": {
     "watch": "concurrently -c 'auto' -n 'core:' 'bun:watch-*'",
-    "watch-browser": "bun build --watch --target=browser --sourcemap=external --external @huggingface/transformers --outdir ./dist ./src/browser.ts",
+    "watch-browser": "bun build --watch --target=browser --sourcemap=external --external @huggingface/transformers --outdir ./dist ./src/index.ts",
     "watch-server-bun": "bun build --watch --target=bun --sourcemap=external --external @huggingface/transformers --outdir ./dist ./src/server.ts",
     "watch-types": "tsc --watch --preserveWatchOutput",
     "build": "bun run build-clean && bun run build-types && bun run build-browser && bun run build-server-bun",
     "build-clean": "rm -fr dist/* tsconfig.tsbuildinfo",
-    "build-browser": "bun build --target=browser --sourcemap=external --external @huggingface/transformers --outdir ./dist ./src/browser*.ts",
-    "build-server-bun": "bun build --target=bun --sourcemap=external --external @huggingface/transformers --outdir ./dist ./src/server*.ts",
+    "build-browser": "bun build --target=browser --sourcemap=external --external @huggingface/transformers --outdir ./dist ./src/index.ts",
+    "build-server-bun": "bun build --target=bun --sourcemap=external --external @huggingface/transformers --outdir ./dist ./src/server.ts",
     "build-types": "tsc",
     "lint": "eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 0",
     "test": "bun test"
   },
-  "module": "dist/server.js",
+  "module": "dist/index.js",
+  "types": "dist/index.d.ts",
   "exports": {
+    ".": {
+      "import": "./dist/index.js",
+      "types": "./dist/index.d.ts"
+    },
     "./browser": {
-      "import": "./dist/browser.js"
+      "import": "./dist/index.js",
+      "types": "./dist/index.d.ts"
     },
     "./server": {
-      "import": "./dist/server.js"
+      "import": "./dist/server.js",
+      "types": "./dist/server.d.ts"
     }
   },
   "files": [
