+++ packages/core/src/index.ts
@@ -9,12 +9,8 @@ export * from "./source/Document";
 export * from "./source/DocumentConverterText";
 export * from "./source/DocumentConverterMarkdown";
 export * from "./task/index";
-export * from "./provider/local-hugging-face/HuggingFaceLocal_TaskRun";
-export * from "./provider/local-media-pipe/MediaPipeLocalTaskRun";
-export * from "./bindings/all_inmemory";
 export * from "./provider/ProviderRegistry";
 export * from "./model/Model";
-export * from "./model/HuggingFaceModel";
 export * from "./model/InMemoryStorage";
 export * from "./storage/taskoutput/TaskOutputRepository";
 export * from "./storage/taskoutput/InMemoryTaskOutputRepository";
