+++ packages/core/src/server.ts
@@ -5,7 +5,7 @@
 //    *   Licensed under the Apache License, Version 2.0 (the "License");           *
 //    *******************************************************************************
 
-export * from "./browser";
+export * from "./index";
 export * from "./storage/taskoutput/SqliteTaskOutputRepository";
 export * from "./storage/taskoutput/PostgresTaskOutputRepository";
 export * from "./storage/taskgraph/SqliteTaskGraphRepository";
@@ -14,4 +14,3 @@ export * from "./job/SqliteJobQueue";
 export * from "./job/SqliteRateLimiter";
 export * from "./job/PostgreSqlJobQueue";
 export * from "./job/PostgreSqlRateLimiter";
-export * from "./bindings/all_sqlite";
