+++ packages/provider-hf-transformers/package.json
@@ -1,0 +1,37 @@
+{
+  "name": "ellmers-provider-hf-transformers",
+  "type": "module",
+  "version": "0.0.1",
+  "description": "Ellmers is a tool for building and running DAG pipelines of AI tasks.",
+  "scripts": {
+    "watch": "concurrently -c 'auto' -n 'tf-meda:' 'bun:watch-*'",
+    "watch-browser": "bun build --watch --target=browser --sourcemap=external --external @huggingface/transformers --external ellmers-core --outdir ./dist ./src/index.ts",
+    "watch-worker": "bun build --watch --target=bun --sourcemap=external --external @huggingface/transformers --external ellmers-core --outdir ./dist ./src/worker.ts",
+    "watch-types": "tsc --watch --preserveWatchOutput",
+    "build": "bun run build-clean && bun run build-types && bun run build-browser && bun run build-worker",
+    "build-clean": "rm -fr dist/* tsconfig.tsbuildinfo",
+    "build-browser": "bun build --target=browser --sourcemap=external --external @huggingface/transformers --external ellmers-core --outdir ./dist ./src/index.ts",
+    "build-worker": "bun build --target=bun --sourcemap=external --external @huggingface/transformers --external ellmers-core --outdir ./dist ./src/worker.ts",
+    "build-types": "tsc",
+    "lint": "eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 0",
+    "test": "bun test"
+  },
+  "module": "dist/index.js",
+  "types": "dist/index.d.ts",
+  "exports": {
+    ".": {
+      "import": "./dist/index.js",
+      "types": "./dist/index.d.ts"
+    },
+    "./worker": {
+      "import": "./dist/worker.js",
+      "types": "./dist/worker.d.ts"
+    }
+  },
+  "files": [
+    "dist"
+  ],
+  "dependencies": {
+    "ellmers-core": "workspace:packages/core"
+  }
+}
