+++ packages/provider-hf-transformers/src/index.ts
@@ -5,4 +5,7 @@
 //    *   Licensed under the Apache License, Version 2.0 (the "License");           *
 //    *******************************************************************************
 
-export * from "./Model";
+export * from "./provider/HuggingFaceLocal_TaskRun";
+export * from "./model/ONNXTransformerJsModel";
+export * from "./bindings/local_hf";
+export * from "./bindings/all_inmemory";
