+++ packages/provider-hf-transformers/tsconfig.json
@@ -1,14 +1,16 @@
 {
   "extends": "../../tsconfig.json",
   "include": ["src/**/*"],
-  "files": ["src/index.ts", "src/server.ts"],
+  "files": ["src/index.ts", "src/worker.ts"],
   "exclude": ["**/*.test.ts", "dist"],
   "compilerOptions": {
     "outDir": "./dist",
     "baseUrl": "./src",
     "rootDir": "./src",
     "paths": {
-      "#/*": ["./src/*"]
+      "#/*": ["./src/*"],
+      "ellmers-core": ["../core/src"]
     }
-  }
+  },
+  "references": [{ "path": "../core" }]
 }
