+++ packages/provider-tf-mediapipe/src/server.ts
@@ -5,4 +5,4 @@
 //    *   Licensed under the Apache License, Version 2.0 (the "License");           *
 //    *******************************************************************************
 
-export * from "./Model";
+export * from ".";
