+++ samples/cli/src/lib.ts
