+++ samples/web/package.json
@@ -21,7 +21,9 @@
     "react-hotkeys-hook": "^4.6.1",
     "react-icons": "^5.4.0",
     "react-resizable-panels": "^2.1.7",
-    "ellmers-core": "workspace:packages/core"
+    "ellmers-core": "workspace:packages/core",
+    "ellmers-provider-hf-transformers": "workspace:packages/provider-hf-transformers",
+    "ellmers-provider-tf-mediapipe": "workspace:packages/provider-tf-mediapipe"
   },
   "devDependencies": {
     "@types/react": "^19.0.4",
