+++ samples/web/src/ConsoleFormatters.tsx
