+++ samples/web/src/GraphStoreStatus.tsx
