+++ samples/web/src/Resize.tsx
