+++ samples/web/src/util.ts
