+++ samples/web/tailwind.config.js
