+++ samples/web/tsconfig.json
@@ -19,9 +19,16 @@
     "noFallthroughCasesInSwitch": true,
 
     "paths": {
-      "ellmers-core": ["./packages/core/"]
+      "ellmers-core": ["../../packages/core/src"],
+      "ellmers-provider-hf-transformers": ["../../packages/provider-hf-transformers/src"],
+      "ellmers-provider-tf-mediapipe": ["../../packages/provider-tf-mediapipe/src"]
     }
   },
-  "include": ["src", "vite.config.js"],
-  "exclude": ["dist"]
+  "include": ["src"],
+  "exclude": ["dist"],
+  "references": [
+    { "path": "../../packages/core" },
+    { "path": "../../packages/provider-hf-transformers" },
+    { "path": "../../packages/provider-tf-mediapipe" }
+  ]
 }
