+++ samples/web/vite.config.js
